split project intro sub-projects
gradle build for multi-project settings
