+++ gradle/wrapper/gradle-wrapper.properties
@@ -1,5 +1,5 @@
 #Wed Apr 22 16:11:33 CEST 2020
-distributionUrl=https\://services.gradle.org/distributions/gradle-6.4.1-all.zip
+distributionUrl=https\://services.gradle.org/distributions/gradle-6.6.1-all.zip
 distributionBase=GRADLE_USER_HOME
 distributionPath=wrapper/dists
 zipStorePath=wrapper/dists
