+++ repository-commons/build.gradle
@@ -1,0 +1,11 @@
+ext {
+    zstdVersion = "1.4.5-6"
+    bcVersion = "1.65"
+}
+
+dependencies {
+    implementation "com.github.luben:zstd-jni:$zstdVersion"
+    implementation "org.bouncycastle:bcprov-jdk15on:$bcVersion"
+
+    compile "org.slf4j:slf4j-api:$slf4jVersion"
+}
