+++ repository-commons/src/main/java/io/aiven/elasticsearch/repositories/metadata/EncryptedRepositoryMetadata.java
