+++ repository-commons/src/main/java/io/aiven/elasticsearch/repositories/security/EncryptionKeyProvider.java
