+++ repository-commons/src/test/java/io/aiven/elasticsearch/repositories/RsaKeyAwareTest.java
