+++ repository-commons/src/test/java/io/aiven/elasticsearch/repositories/metadata/EncryptedRepositoryMetadataTest.java
