+++ repository-commons/src/test/java/io/aiven/elasticsearch/repositories/security/EncryptionKeyProviderTest.java
