+++ repository-commons/src/test/resources/test_gcs_creds.json
