+++ repository-gcs/src/main/java/io.aiven.elasticsearch.repositories.gcs/GcsSettingsProvider.java
