+++ repository-gcs/src/main/resources/plugin-descriptor.properties
