+++ repository-gcs/src/main/resources/plugin-security.policy
