+++ repository-gcs/src/test/java/io.aiven.elasticsearch.repositories.gcs/GcsSettingsProviderTest.java
