+++ repository-gcs/src/test/java/io.aiven.elasticsearch.repositories.gcs/GcsStorageSettingsTest.java
