+++ repository-gcs/src/test/resources/test_gcs_creds.json
@@ -1,0 +1,6 @@
+{
+  "client_id": "test-client-id",
+  "client_secret": "test-client-secret",
+  "refresh_token": "x",
+  "type": "authorized_user"
+}
